+++ scriptconfig/value.py
@@ -1,7 +1,7 @@
 import glob
 import ubelt as ub
 import copy
-from . import smartcast
+from . import smartcast as smartcast_mod
 import re
 
 
@@ -68,6 +68,13 @@ class Value(ub.NiceRepr):
             Indicates that only one of the values in a group should be given on
             the command line. This has no impact on python usage.
 
+        tags (Any):
+            for external program use
+
+    CommandLine:
+        xdoctest -m /home/joncrall/code/scriptconfig/scriptconfig/value.py Value
+        xdoctest -m scriptconfig.value Value
+
     Example:
         >>> self = Value(None, type=float)
         >>> print('self.value = {!r}'.format(self.value))
@@ -83,7 +90,7 @@ class Value(ub.NiceRepr):
     def __init__(self, value=None, type=None, help=None, choices=None,
                  position=None, isflag=False, nargs=None, alias=None,
                  required=False, short_alias=None, group=None,
-                 mutex_group=None):
+                 mutex_group=None, tags=None):
         self.value = None
         self.type = type
         self.alias = alias
@@ -99,10 +106,12 @@ class Value(ub.NiceRepr):
         self.mutex_group = mutex_group
         self.required = required
         self.short_alias = short_alias
+        self.tags = tags
         self.update(value)
 
     def __nice__(self):
-        return '{!r}: {!r}'.format(self.type, self.value)
+        # return '{!r}: {!r}'.format(self.type, self.value)
+        return f'{self.value!r}'
 
     def update(self, value):
         self.value = self.cast(value)
@@ -110,7 +119,7 @@ class Value(ub.NiceRepr):
 
     def cast(self, value):
         if isinstance(value, str):
-            value = smartcast.smartcast(value, self.type)
+            value = smartcast_mod.smartcast(value, self.type)
         return value
 
     def copy(self):
@@ -255,7 +264,7 @@ class PathList(Value):
     def cast(self, value=None):
         if isinstance(value, str):
             paths1 = sorted(glob.glob(ub.expandpath(value)))
-            paths2 = smartcast.smartcast(value)
+            paths2 = smartcast_mod.smartcast(value)
             if paths1:
                 value = paths1
             else:
@@ -412,7 +421,7 @@ def _maker_smart_parse_action(self):
                     template = scfg_object.default[key]
                     if not isinstance(template, Value):
                         # smartcast non-valued params from commandline
-                        value = smartcast.smartcast(value)
+                        value = smartcast_mod.smartcast(value)
                     else:
                         value = template.cast(value)
                     return value